refactor: move Order class from menu to order package
